+++ Assets/Nova/Sources/Core/Restoration/GameStateCheckpoint.cs
