+++ Assets/Nova/Sources/Core/Restoration/ReachedDialogueData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 776b3975fc5cc0249a4c912815850ada
+guid: 9693bfe057cd98d48b5827318bb26975
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
